+++ library/ui/src/main/java/com/google/android/exoplayer2/ui/spherical/TouchTracker.java
@@ -44,7 +44,7 @@ import androidx.annotation.Nullable;
  * a nicer UI. An even more advanced UI would reproject the user's touch point into 3D and drag the
  * Mesh as the user moves their finger. However, that requires quaternion interpolation.
  */
-/* package */ class TouchTracker extends GestureDetector.SimpleOnGestureListener
+/* package */ final class TouchTracker extends GestureDetector.SimpleOnGestureListener
     implements View.OnTouchListener, OrientationListener.Listener {
 
   /* package */ interface Listener {