+++ RELEASENOTES.md
@@ -8,6 +8,8 @@
     *   Add builder for `PlayerNotificationManager`.
 *   Library restructuring:
     *   `DebugTextViewHelper` moved from `ui` package to `util` package.
+    *   Spherical UI components moved from `video.spherical` package to
+        `ui.spherical` package, and made package private.
 *   Remove deprecated symbols:
     *   Remove `Player.DefaultEventListener`. Use `Player.EventListener`
         instead.